+++ packages/isaacscript-common/src/features/collectibleItemPoolType.ts
@@ -1,5 +1,5 @@
 // The item pool type of a collectible is not stored on the collectible. Thus, we scan for incoming
-// item pool types in the PreGetCollectible callback, and then assume that the next spawned
+// item pool types in the `PRE_GET_COLLECTIBLE` callback, and then assume that the next spawned
 // collectible has this item pool type.
 
 import {
@@ -45,8 +45,8 @@ function postPickupInitCollectible(pickup: EntityPickup) {
 
 /**
  * Helper function to get the item pool type that a given collectible came from. Since there is no
- * native method in the API to get this, we listen in the PreGetCollectible callback for item pool
- * types, and then assume that the next spawned collectible will match.
+ * native method in the API to get this, we listen in the `PRE_GET_COLLECTIBLE` callback for item
+ * pool types, and then assume that the next spawned collectible will match.
  */
 export function getCollectibleItemPoolType(
   collectible: EntityPickup,